requirements.txt
pandas
pymongo
google-cloud-bigquery
python-dotenv